requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
oauthlib==2.0.2
PyJWT==1.5.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
docopt==0.6.2
schema==0.6.6
faker==0.7.18
msgpack-python==0.4.8
aiohttp==2.3.6
